#106 Reorganized functions in files
